* Small improvement. Saves a byte.

+++ NewRemoteSwitch/NewRemoteTransmitter.cpp
@@ -18,7 +18,7 @@ NewRemoteTransmitter::NewRemoteTransmitter(unsigned long address, unsigned short
 }
 
 void NewRemoteTransmitter::sendGroup(boolean switchOn) {
-	for (int i = _repeats; i >= 0; i--) {
+	for (unsigned short i = _repeats; i >= 0; i--) {
 		_sendStartPulse();
 
 		_sendAddress();
@@ -37,7 +37,7 @@ void NewRemoteTransmitter::sendGroup(boolean switchOn) {
 }
 
 void NewRemoteTransmitter::sendUnit(unsigned short unit, boolean switchOn) {
-	for (int i = _repeats; i >= 0; i--) {
+	for (unsigned short i = _repeats; i >= 0; i--) {
 		_sendStartPulse();
 
 		_sendAddress();
@@ -55,7 +55,7 @@ void NewRemoteTransmitter::sendUnit(unsigned short unit, boolean switchOn) {
 }
 
 void NewRemoteTransmitter::sendDim(unsigned short unit, unsigned short dimLevel) {
-	for (int i = _repeats; i >= 0; i--) {
+	for (unsigned short i = _repeats; i >= 0; i--) {
 		_sendStartPulse();
 
 		_sendAddress();
